+++ app/pb_kits/playbook/pb_avatar_action_button/docs/_avatar_action_button_default.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { AvatarActionButton } from '../../'
+
+const AvatarActionButtonDefault = () => (
+  <div className="pb--doc-demo-row">
+    <AvatarActionButton
+        imageUrl="https://randomuser.me/api/portraits/women/8.jpg"
+        name="Sophia Carden"
+    />
+  </div>
+)
+
+export default AvatarActionButtonDefault